Move cropping and texture and font rendering into CanvasRenderer
