+++ engine/src/main/java/org/terasology/rendering/nui/internal/LwjglCanvasRenderer.java
@@ -22,17 +22,35 @@ import org.terasology.asset.Assets;
 import org.terasology.math.AABB;
 import org.terasology.math.MatrixUtils;
 import org.terasology.math.Quat4fUtil;
+import org.terasology.math.Rect2f;
 import org.terasology.math.Rect2i;
 import org.terasology.math.Vector2i;
+import org.terasology.rendering.assets.font.Font;
 import org.terasology.rendering.assets.material.Material;
 import org.terasology.rendering.assets.mesh.Mesh;
 import org.terasology.rendering.assets.shader.ShaderProgramFeature;
+import org.terasology.rendering.assets.texture.TextureRegion;
 import org.terasology.rendering.nui.Color;
+import org.terasology.rendering.nui.HorizontalAlign;
+import org.terasology.rendering.nui.ScaleMode;
+import org.terasology.rendering.nui.TextLineBuilder;
+import org.terasology.rendering.nui.VerticalAlign;
+
+import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
 
 import javax.vecmath.Matrix4f;
 import javax.vecmath.Quat4f;
+import javax.vecmath.Vector2f;
 import javax.vecmath.Vector3f;
+import javax.vecmath.Vector4f;
+
 import java.nio.FloatBuffer;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 
 import static org.lwjgl.opengl.GL11.GL_BLEND;
 import static org.lwjgl.opengl.GL11.GL_DEPTH_TEST;
@@ -63,6 +81,12 @@ public class LwjglCanvasRenderer implements CanvasRenderer {
     private Mesh billboard = Assets.getMesh("engine:UIBillboard");
     private Line line = new Line();
 
+    private Material textureMat = Assets.getMaterial("engine:UITexture");
+
+    // Text mesh caching
+    private Map<TextCacheKey, Map<Material, Mesh>> cachedText = Maps.newLinkedHashMap();
+    private Set<TextCacheKey> usedText = Sets.newHashSet();
+
     @Override
     public void preRender() {
         glDisable(GL_DEPTH_TEST);
@@ -87,6 +111,18 @@ public class LwjglCanvasRenderer implements CanvasRenderer {
 
     @Override
     public void postRender() {
+        Iterator<Map.Entry<TextCacheKey, Map<Material, Mesh>>> textIterator = cachedText.entrySet().iterator();
+        while (textIterator.hasNext()) {
+            Map.Entry<TextCacheKey, Map<Material, Mesh>> entry = textIterator.next();
+            if (!usedText.contains(entry.getKey())) {
+                for (Mesh mesh : entry.getValue().values()) {
+                    Assets.dispose(mesh);
+                }
+                textIterator.remove();
+            }
+        }
+        usedText.clear();
+
         glMatrixMode(GL_MODELVIEW);
         glPopMatrix();
         glMatrixMode(GL_PROJECTION);
@@ -160,4 +196,92 @@ public class LwjglCanvasRenderer implements CanvasRenderer {
         line.draw(sx, sy, ex, ey, 2, color, color, 0);
     }
 
+    @Override
+    public void crop(Rect2i cropRegion) {
+        textureMat.setFloat4("croppingBoundaries", cropRegion.minX(), cropRegion.maxX() + 1, cropRegion.minY(), cropRegion.maxY() + 1);
+    }
+
+    public void drawTexture(TextureRegion texture, Color color, ScaleMode mode, Rect2i absoluteRegion,
+                            float ux, float uy, float uw, float uh, float alpha) {
+        Vector2f scale = mode.scaleForRegion(absoluteRegion, texture.getWidth(), texture.getHeight());
+        Rect2f textureArea = texture.getRegion();
+        textureMat.setFloat2("scale", scale);
+        textureMat.setFloat2("offset",
+                absoluteRegion.minX() + 0.5f * (absoluteRegion.width() - scale.x),
+                absoluteRegion.minY() + 0.5f * (absoluteRegion.height() - scale.y));
+        textureMat.setFloat2("texOffset", textureArea.minX() + ux * textureArea.width(), textureArea.minY() + uy * textureArea.height());
+        textureMat.setFloat2("texSize", uw * textureArea.width(), uh * textureArea.height());
+        textureMat.setTexture("texture", texture.getTexture());
+        textureMat.setFloat4("color", color.rf(), color.gf(), color.bf(), color.af() * alpha);
+        textureMat.bindTextures();
+        billboard.render();
+    }
+
+    public void drawText(String text, Font font, HorizontalAlign hAlign, VerticalAlign vAlign, Rect2i absoluteRegion, Rect2i cropRegion,
+                         Color color, Color shadowColor, float alpha) {
+        TextCacheKey key = new TextCacheKey(text, font, absoluteRegion.width(), hAlign);
+        usedText.add(key);
+        Map<Material, Mesh> fontMesh = cachedText.get(key);
+        List<String> lines = TextLineBuilder.getLines(font, text, absoluteRegion.width());
+        if (fontMesh == null) {
+            fontMesh = font.createTextMesh(lines, absoluteRegion.width(), hAlign);
+            cachedText.put(key, fontMesh);
+        }
+
+        Vector2i offset = new Vector2i(absoluteRegion.minX(), absoluteRegion.minY());
+        offset.y += vAlign.getOffset(lines.size() * font.getLineHeight(), absoluteRegion.height());
+
+        for (Map.Entry<Material, Mesh> entry : fontMesh.entrySet()) {
+            entry.getKey().bindTextures();
+            entry.getKey().setFloat4("croppingBoundaries", cropRegion.minX(), cropRegion.maxX() + 1, cropRegion.minY(), cropRegion.maxY() + 1);
+            if (shadowColor.a() != 0) {
+                entry.getKey().setFloat2("offset", offset.x + 1, offset.y + 1);
+                Vector4f shadowValues = shadowColor.toVector4f();
+                shadowValues.w *= alpha;
+                entry.getKey().setFloat4("color", shadowValues);
+                entry.getValue().render();
+            }
+
+            entry.getKey().setFloat2("offset", offset.x, offset.y);
+            Vector4f colorValues = color.toVector4f();
+            colorValues.w *= alpha;
+            entry.getKey().setFloat4("color", colorValues);
+            entry.getValue().render();
+        }
+    }
+
+    /**
+     * A key that identifies an entry in the text cache. It contains the elements that affect the generation of mesh for text rendering.
+     */
+    private static class TextCacheKey {
+        private String text;
+        private Font font;
+        private int width;
+        private HorizontalAlign alignment;
+
+        public TextCacheKey(String text, Font font, int maxWidth, HorizontalAlign alignment) {
+            this.text = text;
+            this.font = font;
+            this.width = maxWidth;
+            this.alignment = alignment;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (this == obj) {
+                return true;
+            }
+            if (obj instanceof TextCacheKey) {
+                TextCacheKey other = (TextCacheKey) obj;
+                return Objects.equals(text, other.text) && Objects.equals(font, other.font)
+                       && Objects.equals(width, other.width) && Objects.equals(alignment, other.alignment);
+            }
+            return false;
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(text, font, width, alignment);
+        }
+    }
 }
